Extract worker tests from one large file Moved into logical smaller units
Change-type: minor
